rename xiaomi to xiaomi_aqara
