+++ homeassistant/components/sensor/xiaomi_aqara.py
@@ -1,7 +1,7 @@
-"""Support for Xiaomi sensors."""
+"""Support for Xiaomi aqara sensors."""
 import logging
 
-from homeassistant.components.xiaomi import (PY_XIAOMI_GATEWAY, XiaomiDevice)
+from homeassistant.components.xiaomi_aqara import (PY_XIAOMI_GATEWAY, XiaomiDevice)
 from homeassistant.const import TEMP_CELSIUS
 
 _LOGGER = logging.getLogger(__name__)
